requirements.txt
flask
flask-sqlalchemy
flask-smorest
python-dotenv
marshmallow
flask-jwt-extended
passlib
flask-jwt-extended
